+++ src/Rule.php
@@ -554,7 +554,7 @@ class Rule extends CommonDBTM
     /**
      * @since 0.84
      *
-     * @return string
+     * @return class-string<RuleCollection>
      **/
     public function getCollectionClassName()
     {
@@ -1985,10 +1985,65 @@ JS
             );
         }
 
-        // Before adding, add the ranking of the new rule
-        $input["ranking"] ??= $this->getNextRanking($input['sub_type']);
+        // Before adding, add the ranking of the new rule to be handled via the moveRule method
+        $valid_ranking = isset($input['ranking']) && is_numeric($input['ranking']) && $input['ranking'] >= 0;
+        $input["_ranking"] = (int) ($valid_ranking ? $input['ranking'] : $this->getNextRanking($input['sub_type']));
+        unset($input['ranking']);
 
         return $input;
+    }
+
+    public function prepareInputForUpdate($input)
+    {
+        $input = parent::prepareInputForUpdate($input);
+
+        if ($input !== false) {
+            if (isset($input['ranking'])) {
+                // Could be set this way from the API.
+                // In this case, we should use moveRule rather than updating directly.
+                $input["_ranking"] = $input['ranking'];
+                unset($input['ranking']);
+            } else if (isset($input['_ranking'])) {
+                // Set this way from RuleCollection::moveRule to avoid infinite loop
+                $input['ranking'] = $input['_ranking'];
+                unset($input['_ranking']);
+            }
+        }
+        return $input;
+    }
+
+    public function post_addItem()
+    {
+        parent::post_addItem();
+        $this->handleRankChange(true);
+    }
+
+    public function post_updateItem($history = true)
+    {
+        parent::post_updateItem($history);
+        $this->handleRankChange();
+    }
+
+    /**
+     * Handles any rank change from the API or another source except {@link RuleCollection::moveRule()}
+     * by moving the rule rather than directly setting the rank to handle the other rules and avoid rules with the same rank.
+     * @return void
+     */
+    private function handleRankChange($new_rule = false)
+    {
+        if (isset($this->input['_ranking'])) {
+            if (isset($this->fields['ranking']) && (int) $this->input['_ranking'] === (int) $this->fields['ranking']) {
+                // No change in ranking, nothing to do.
+                return;
+            }
+            // Ensure we always use the right rule class in case the original object was instantiated as the base class 'Rule'.
+            $rule_class = $this->fields['sub_type'] ?? static::class;
+            $rule = new $rule_class();
+            $collection_class = $rule->getCollectionClassName();
+            $collection = new $collection_class();
+            $collection->moveRule($this->fields['id'], 0, $this->input['_ranking'], $new_rule);
+            $this->getFromDB($this->fields['id']);
+        }
     }
 
     /**
@@ -2010,7 +2065,7 @@ JS
 
         if (count($iterator)) {
             $data = $iterator->current();
-            return $data["rank"] + 1;
+            return $data["rank"] === null ? 0 : $data['rank'] + 1;
         }
         return 0;
     }